+++ SBO-Lib/ChangeLog.md
@@ -6,6 +6,7 @@
     * Thanks to fsLeg.
   * Readme.pm: Parse SlackBuilds for useradd and groupadd commands first
   * Readme.pm: Ask about options before asking about users and groups
+  * Readme.pm: Handle SlackBuilds with variable user and group names and UID/GID
 
 #### 4.1.2 - 2025-11-21
   * sbocheck: Perl test properly checks perl-involved shared objects